# api/openapi.yaml
openapi: 3.0.3
info:
  title: Worksheet Checker API
  version: 0.1.0
...
paths:
...
  /classes:
    get:
      operationId: listClasses
      tags:
        - classes
      summary: List Classes owned by the authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassList"
    post:
      operationId: createClass
      tags:
        - classes
      summary: Create a Class owned by the authenticated user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
  /classes/{classId}:
    get:
      operationId: getClass
      tags:
        - classes
      summary: Get a Class owned by the authenticated user
      security:
        - bearerAuth: []
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
        "404":
          description: Not found
    put:
      operationId: updateClass
      tags:
        - classes
      summary: Update a Class owned by the authenticated user
      security:
        - bearerAuth: []
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClassRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
        "404":
          description: Not found
    delete:
      operationId: deleteClass
      tags:
        - classes
      summary: Delete a Class owned by the authenticated user
      security:
        - bearerAuth: []
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /classes/{classId}/subjects:
    get:
      operationId: listClassSubjects
      tags:
        - classes
      summary: List Subjects selected for a Class
      security:
        - bearerAuth: []
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubjectList"
        "404":
          description: Not found
    post:
      operationId: addClassSubject
      tags:
        - classes
      summary: Select a Subject for a Class
      security:
        - bearerAuth: []
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddClassSubjectRequest"
      responses:
        "201":
          description: Selected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubjectList"
        "404":
          description: Not found
  /classes/{classId}/subjects/{subjectId}:
    delete:
      operationId: removeClassSubject
      tags:
        - classes
      summary: Remove a Subject selection from a Class
      security:
        - bearerAuth: []
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: subjectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Removed
        "404":
          description: Not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Class:
      type: object
      required:
        - id
        - name
        - created_by
        - created_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
    ClassList:
      type: object
      required:
        - classes
      properties:
        classes:
          type: array
          items:
            $ref: "#/components/schemas/Class"
    CreateClassRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
    UpdateClassRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
    Subject:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    SubjectList:
      type: object
      required:
        - subjects
      properties:
        subjects:
          type: array
          items:
            $ref: "#/components/schemas/Subject"
    AddClassSubjectRequest:
      type: object
      required:
        - subject_id
      properties:
        subject_id:
          type: string
          format: uuid
...
servers:
  - url: http://localhost:8080
    description: Local