# backend/gen/api/classes/schema.gen.go
// Package classes_gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package classes_gen

import (
	"fmt"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// AddClassSubjectRequest defines model for AddClassSubjectRequest.
type AddClassSubjectRequest struct {
	SubjectId openapi_types.UUID `json:"subject_id"`
}

// Class defines model for Class.
type Class struct {
	CreatedAt time.Time          `json:"created_at"`
	CreatedBy openapi_types.UUID `json:"created_by"`
	Id        openapi_types.UUID `json:"id"`
	Name      string             `json:"name"`
}

// ClassList defines model for ClassList.
type ClassList struct {
	Classes []Class `json:"classes"`
}

// CreateClassRequest defines model for CreateClassRequest.
type CreateClassRequest struct {
	Name string `json:"name"`
}

// Subject defines model for Subject.
type Subject struct {
	Id   openapi_types.UUID `json:"id"`
	Name string             `json:"name"`
}

// SubjectList defines model for SubjectList.
type SubjectList struct {
	Subjects []Subject `json:"subjects"`
}

// UpdateClassRequest defines model for UpdateClassRequest.
type UpdateClassRequest struct {
	Name string `json:"name"`
}

// CreateClassJSONRequestBody defines body for CreateClass for application/json ContentType.
type CreateClassJSONRequestBody = CreateClassRequest

// UpdateClassJSONRequestBody defines body for UpdateClass for application/json ContentType.
type UpdateClassJSONRequestBody = UpdateClassRequest

// AddClassSubjectJSONRequestBody defines body for AddClassSubject for application/json ContentType.
type AddClassSubjectJSONRequestBody = AddClassSubjectRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// List Classes owned by the authenticated user
	// (GET /classes)
	ListClasses(c *gin.Context)
	// Create a Class owned by the authenticated user
	// (POST /classes)
	CreateClass(c *gin.Context)
	// Delete a Class owned by the authenticated user
	// (DELETE /classes/{classId})
	DeleteClass(c *gin.Context, classId openapi_types.UUID)
	// Get a Class owned by the authenticated user
	// (GET /classes/{classId})
	GetClass(c *gin.Context, classId openapi_types.UUID)
	// Update a Class owned by the authenticated user
	// (PUT /classes/{classId})
	UpdateClass(c *gin.Context, classId openapi_types.UUID)
	// List Subjects selected for a Class
	// (GET /classes/{classId}/subjects)
	ListClassSubjects(c *gin.Context, classId openapi_types.UUID)
	// Select a Subject for a Class
	// (POST /classes/{classId}/subjects)
	AddClassSubject(c *gin.Context, classId openapi_types.UUID)
	// Remove a Subject selection from a Class
	// (DELETE /classes/{classId}/subjects/{subjectId})
	RemoveClassSubject(c *gin.Context, classId openapi_types.UUID, subjectId openapi_types.UUID)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// ListClasses operation middleware
func (siw *ServerInterfaceWrapper) ListClasses(c *gin.Context) {

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.ListClasses(c)
}

// CreateClass operation middleware
func (siw *ServerInterfaceWrapper) CreateClass(c *gin.Context) {

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.CreateClass(c)
}

// DeleteClass operation middleware
func (siw *ServerInterfaceWrapper) DeleteClass(c *gin.Context) {

	var err error

	// ------------- Path parameter "classId" -------------
	var classId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "classId", c.Param("classId"), &classId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter classId: %w", err), http.StatusBadRequest)
		return
	}

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.DeleteClass(c, classId)
}

// GetClass operation middleware
func (siw *ServerInterfaceWrapper) GetClass(c *gin.Context) {

	var err error

	// ------------- Path parameter "classId" -------------
	var classId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "classId", c.Param("classId"), &classId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter classId: %w", err), http.StatusBadRequest)
		return
	}

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetClass(c, classId)
}

// UpdateClass operation middleware
func (siw *ServerInterfaceWrapper) UpdateClass(c *gin.Context) {

	var err error

	// ------------- Path parameter "classId" -------------
	var classId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "classId", c.Param("classId"), &classId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter classId: %w", err), http.StatusBadRequest)
		return
	}

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.UpdateClass(c, classId)
}

// ListClassSubjects operation middleware
func (siw *ServerInterfaceWrapper) ListClassSubjects(c *gin.Context) {

	var err error

	// ------------- Path parameter "classId" -------------
	var classId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "classId", c.Param("classId"), &classId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter classId: %w", err), http.StatusBadRequest)
		return
	}

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.ListClassSubjects(c, classId)
}

// AddClassSubject operation middleware
func (siw *ServerInterfaceWrapper) AddClassSubject(c *gin.Context) {

	var err error

	// ------------- Path parameter "classId" -------------
	var classId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "classId", c.Param("classId"), &classId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter classId: %w", err), http.StatusBadRequest)
		return
	}

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.AddClassSubject(c, classId)
}

// RemoveClassSubject operation middleware
func (siw *ServerInterfaceWrapper) RemoveClassSubject(c *gin.Context) {

	var err error

	// ------------- Path parameter "classId" -------------
	var classId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "classId", c.Param("classId"), &classId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter classId: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Path parameter "subjectId" -------------
	var subjectId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "subjectId", c.Param("subjectId"), &subjectId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter subjectId: %w", err), http.StatusBadRequest)
		return
	}

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.RemoveClassSubject(c, classId, subjectId)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.GET(options.BaseURL+"/classes", wrapper.ListClasses)
	router.POST(options.BaseURL+"/classes", wrapper.CreateClass)
	router.DELETE(options.BaseURL+"/classes/:classId", wrapper.DeleteClass)
	router.GET(options.BaseURL+"/classes/:classId", wrapper.GetClass)
	router.PUT(options.BaseURL+"/classes/:classId", wrapper.UpdateClass)
	router.GET(options.BaseURL+"/classes/:classId/subjects", wrapper.ListClassSubjects)
	router.POST(options.BaseURL+"/classes/:classId/subjects", wrapper.AddClassSubject)
	router.DELETE(options.BaseURL+"/classes/:classId/subjects/:subjectId", wrapper.RemoveClassSubject)
}