# backend/internal/database/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package database

import (
	"time"

	"github.com/google/uuid"
)

type AuthToken struct {
	ID        uuid.UUID
	UserID    uuid.UUID
	Kind      string
	TokenHash string
	ExpiresAt time.Time
	UsedAt    *time.Time
	CreatedAt time.Time
}

type Class struct {
	ID        uuid.UUID
	Name      string
	CreatedBy uuid.UUID
	CreatedAt time.Time
	UpdatedAt time.Time
}

type ClassSubject struct {
	ClassID   uuid.UUID
	SubjectID uuid.UUID
	CreatedAt time.Time
}

type RefreshToken struct {
	ID        uuid.UUID
	UserID    uuid.UUID
	TokenHash string
	ExpiresAt time.Time
	RevokedAt *time.Time
	CreatedAt time.Time
}

type Student struct {
	ID         uuid.UUID
	ClassID    uuid.UUID
	Name       string
	RollNumber string
	CreatedAt  time.Time
	UpdatedAt  time.Time
}

type Subject struct {
	ID        uuid.UUID
	Name      string
	CreatedAt time.Time
}

type User struct {
	ID           uuid.UUID
	Email        string
	PasswordHash string
	Name         string
	Role         string
	Verified     bool
	CreatedAt    time.Time
	UpdatedAt    time.Time
}